refactored and introduced AbstractS3RetryStore to reduce code duplication
