+++ interlok-aws2-s3/build.gradle
@@ -10,6 +10,8 @@ ext {
 }
 
 dependencies {
+  implementation project(':interlok-aws-s3')
+
   api ("software.amazon.awssdk:s3:$awsSDKVersion") {
     exclude group: "com.fasterxml.jackson.core", module: "jackson-databind"
     exclude group: "com.fasterxml.jackson.core", module: "jackson-core"
